refactor(attachment-url): move URL validation and download logic to a separate utility module
